Separate implementation from interface
