+++ cpp/include/kvikio/shim/cuda.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2024, NVIDIA CORPORATION.
+ * Copyright (c) 2022-2025, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +15,7 @@
  */
 #pragma once
 
+#include <kvikio/shim/cuda.hpp>
 #include <kvikio/shim/cuda_h_wrapper.hpp>
 #include <kvikio/shim/utils.hpp>
 
@@ -50,46 +51,13 @@ class cudaAPI {
   decltype(cuStreamDestroy)* StreamDestroy{nullptr};
 
  private:
-#ifdef KVIKIO_CUDA_FOUND
-  cudaAPI()
-  {
-    void* lib = load_library("libcuda.so.1");
-    // Notice, the API version loaded must match the version used downstream. That is,
-    // if a project uses the `_v2` CUDA Driver API or the newest Runtime API, the symbols
-    // loaded should also be the `_v2` symbols. Thus, we use KVIKIO_STRINGIFY() to get
-    // the name of the symbol through cude.h.
-    get_symbol(MemHostAlloc, lib, KVIKIO_STRINGIFY(cuMemHostAlloc));
-    get_symbol(MemFreeHost, lib, KVIKIO_STRINGIFY(cuMemFreeHost));
-    get_symbol(MemcpyHtoDAsync, lib, KVIKIO_STRINGIFY(cuMemcpyHtoDAsync));
-    get_symbol(MemcpyDtoHAsync, lib, KVIKIO_STRINGIFY(cuMemcpyDtoHAsync));
-    get_symbol(PointerGetAttribute, lib, KVIKIO_STRINGIFY(cuPointerGetAttribute));
-    get_symbol(PointerGetAttributes, lib, KVIKIO_STRINGIFY(cuPointerGetAttributes));
-    get_symbol(CtxPushCurrent, lib, KVIKIO_STRINGIFY(cuCtxPushCurrent));
-    get_symbol(CtxPopCurrent, lib, KVIKIO_STRINGIFY(cuCtxPopCurrent));
-    get_symbol(CtxGetCurrent, lib, KVIKIO_STRINGIFY(cuCtxGetCurrent));
-    get_symbol(MemGetAddressRange, lib, KVIKIO_STRINGIFY(cuMemGetAddressRange));
-    get_symbol(GetErrorName, lib, KVIKIO_STRINGIFY(cuGetErrorName));
-    get_symbol(GetErrorString, lib, KVIKIO_STRINGIFY(cuGetErrorString));
-    get_symbol(DeviceGet, lib, KVIKIO_STRINGIFY(cuDeviceGet));
-    get_symbol(DevicePrimaryCtxRetain, lib, KVIKIO_STRINGIFY(cuDevicePrimaryCtxRetain));
-    get_symbol(DevicePrimaryCtxRelease, lib, KVIKIO_STRINGIFY(cuDevicePrimaryCtxRelease));
-    get_symbol(StreamSynchronize, lib, KVIKIO_STRINGIFY(cuStreamSynchronize));
-    get_symbol(StreamCreate, lib, KVIKIO_STRINGIFY(cuStreamCreate));
-    get_symbol(StreamDestroy, lib, KVIKIO_STRINGIFY(cuStreamDestroy));
-  }
-#else
-  cudaAPI() { throw std::runtime_error("KvikIO not compiled with CUDA support"); }
-#endif
+  cudaAPI();
 
  public:
   cudaAPI(cudaAPI const&)        = delete;
   void operator=(cudaAPI const&) = delete;
 
-  KVIKIO_EXPORT static cudaAPI& instance()
-  {
-    static cudaAPI _instance;
-    return _instance;
-  }
+  KVIKIO_EXPORT static cudaAPI& instance();
 };
 
 /**
@@ -100,15 +68,7 @@ class cudaAPI {
  * @return The boolean answer
  */
 #ifdef KVIKIO_CUDA_FOUND
-inline bool is_cuda_available()
-{
-  try {
-    cudaAPI::instance();
-  } catch (const std::runtime_error&) {
-    return false;
-  }
-  return true;
-}
+bool is_cuda_available();
 #else
 constexpr bool is_cuda_available() { return false; }
 #endif
