+++ cpp/src/cufile/config.cpp
@@ -13,19 +13,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#pragma once
 
-#include <kvikio/utils.hpp>
+#include <cstdlib>
+#include <filesystem>
+#include <string>
+
+#include <kvikio/cufile/config.hpp>
 
 namespace kvikio {
+namespace detail {
 
-/**
- * @brief Get the filepath to cuFile's config file (`cufile.json`) or the empty string
- *
- * This lookup is cached.
- *
- * @return The filepath to the cufile.json file or the empty string if it isn't found.
- */
-[[nodiscard]] KVIKIO_EXPORT const std::string& config_path();
+[[nodiscard]] inline const char* lookup_config_path()
+{
+  const char* env = std::getenv("CUFILE_ENV_PATH_JSON");
+  if (env != nullptr && std::filesystem::exists(env)) { return env; }
+  if (std::filesystem::exists("/etc/cufile.json")) { return "/etc/cufile.json"; }
+  return "";
+}
+
+}  // namespace detail
+
+const std::string& config_path()
+{
+  static const std::string ret = detail::lookup_config_path();
+  return ret;
+}
 
 }  // namespace kvikio
